Restructure of lightclient modules
