+++ raiden/raiden_service.py
@@ -31,26 +31,23 @@ from raiden.exceptions import (
     RaidenRecoverableError,
     RaidenUnrecoverableError,
     InvalidPaymentIdentifier)
-from raiden.lightclient.light_client_message_handler import LightClientMessageHandler
-from raiden.lightclient.lightclientmessages.light_client_protocol_message import LightClientProtocolMessageType
-from raiden.message_event_convertor import message_from_sendevent
+from raiden.lightclient.handlers.light_client_message_handler import LightClientMessageHandler
+from raiden.lightclient.models.light_client_protocol_message import LightClientProtocolMessageType
 from raiden.messages import (
     LockedTransfer,
     Message,
-    RequestMonitoring,
     SignedMessage,
     RevealSecret, Unlock, Delivered, SecretRequest, Processed)
 from raiden.network.blockchain_service import BlockChainService
 from raiden.network.proxies.secret_registry import SecretRegistry
 from raiden.network.proxies.service_registry import ServiceRegistry
 from raiden.network.proxies.token_network_registry import TokenNetworkRegistry
-from raiden.settings import MEDIATION_FEE, MONITORING_MIN_CAPACITY, MONITORING_REWARD
+from raiden.settings import MEDIATION_FEE
 from raiden.storage import serialize, sqlite, wal
 from raiden.tasks import AlarmTask
-from raiden.transfer import channel, node, views
+from raiden.transfer import node, views
 from raiden.transfer.architecture import Event as RaidenEvent, StateChange
-from raiden.transfer.mediated_transfer.events import SendLockedTransfer, SendLockedTransferLight, \
-    CHANNEL_IDENTIFIER_GLOBAL_QUEUE
+from raiden.transfer.mediated_transfer.events import SendLockedTransfer, SendLockedTransferLight
 
 from raiden.transfer.mediated_transfer.state import (
     TransferDescriptionWithSecretState,
@@ -76,7 +73,7 @@ from raiden.transfer.state_change import (
     Block,
     ContractReceiveNewPaymentNetwork,
 )
-from raiden.utils import create_default_identifier, lpex, pex, random_secret, sha3, to_rdn
+from raiden.utils import create_default_identifier, lpex, pex, random_secret, sha3
 from raiden.utils.runnable import Runnable
 from raiden.utils.signer import LocalSigner, Signer
 from raiden.utils.typing import (
@@ -93,13 +90,11 @@ from raiden.utils.typing import (
     TargetAddress,
     TokenNetworkAddress,
     TokenNetworkID,
-    PaymentHashInvoice,
-    AddressHex)
+    PaymentHashInvoice)
 
 from raiden.utils.upgrades import UpgradeManager
 from raiden_contracts.contract_manager import ContractManager
-from eth_utils import to_canonical_address, to_checksum_address, encode_hex
-from raiden.lightclient.light_client_service import LightClientService
+from eth_utils import to_canonical_address, to_checksum_address
 
 log = structlog.get_logger(__name__)  # pylint: disable=invalid-name
 StatusesDict = Dict[TargetAddress, Dict[PaymentID, "PaymentStatus"]]
